Seção 9: DTOs e Data Annotation

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using ProEventos.Domain;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -44,7 +45,7 @@ namespace ProEventos.API.Controllers
             try
             {
                  var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                 if (evento == null) return NotFound("Eventos por Id não encontrados.");
+                 if (evento == null) return NoContent();
 
                  return Ok(evento);
             }
@@ -61,7 +62,7 @@ namespace ProEventos.API.Controllers
             try
             {
                  var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                 if (evento == null) return NotFound("Eventos por tema não encontrados.");
+                 if (evento == null) return NoContent();
 
                  return Ok(evento);
             }
@@ -73,12 +74,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                  var evento = await _eventoService.AddEventos(model);
-                 if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                 if (evento == null) return NoContent();
 
                  return Ok(evento);
             }
@@ -90,12 +91,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
              try
             {
                  var evento = await _eventoService.UpdateEvento(id, model);
-                 if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                 if (evento == null) return NoContent();
 
                  return Ok(evento);
             }
@@ -111,9 +112,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
-                 return await _eventoService.DeleteEvento(id) ?
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
+                return await _eventoService.DeleteEvento(id) ?
                         Ok("Deletado.") :
-                        BadRequest("Evento não deletado.");
+                        throw new Exception("Ocorreu um problema não específicado ao tentar deletar Evento.");
                  
             }
             catch (Exception ex)
